修改目录结构，添加静态分析器
