+++ snoop/config.json
