+++ snoop/config_template.json
@@ -1,0 +1,25 @@
+{
+    "snoop_cpu": "top",   
+    "cpu_output_file": "cpu.csv",
+    "snoop_mem": "bcc",
+    "mem_output_file": "mem.csv",
+    "probes":{						
+        "event_name":["./test/high_cpu_test:func2"],
+        "output_file":"tmp.csv",
+        "spid": 10086
+    },
+    "trace": {						
+        "tracee_name":["./test/high_cpu_test:func2"],
+        "output_file":"trace_output",			
+        "enter_msg_format":["Enter test"],
+        "return_msg_format":["Return test"],
+        "spid": 10086
+    },
+    "snoop_network": "bcc",
+    "network_output_file": "net.csv",
+    "snoop_syscall": "bcc",
+    "syscall_output_file": "syscall.csv",
+    "interval": 3,
+    "trace_multiprocess": true,
+    "show_all_threads": false
+}
