--- snoop/new/run.sh
@@ -1,4 +1,0 @@
-#! /bin/bash
-sudo ./top_snoop.py \
--c "./fork_test" \
---configure_file ./config.json
