+++ snoop/tracer.py
