+++ snoop/utils.py
@@ -1,4 +1,7 @@
 import subprocess
+import psutil
+import queue
+import time
 
 def pid_to_comm(pid):
     """根据pid查找/proc/pid/comm获取comm
@@ -11,7 +14,7 @@ def pid_to_comm(pid):
     """
     try:
         comm = open("/proc/%d/comm" % pid, "r").read()
-        return comm
+        return comm.replace("\n", "")
     except IOError:
         return str(pid)
 
@@ -52,3 +55,35 @@ def run_command(command):
     """
     p = subprocess.Popen(command.split())
     return p
+
+def bfs_get_procs(snoop_pid):
+    """使用广度优先搜索获取snoop_pid进程下的所有子进程、孙子进程...
+
+    Args:
+        snoop_pid (int): 根节点进程pid
+
+    Returns:
+        list: 进程树下所有进程的pid
+    """
+    proc = psutil.Process(snoop_pid)
+    proc_queue = queue.Queue()
+    proc_queue.put(proc)
+    pid_list = []
+    while not proc_queue.empty():
+        proc = proc_queue.get()
+        pid_list.append(proc.pid)
+        list(map(proc_queue.put, proc.children()))
+    
+    return pid_list
+
+def get_delta():
+    """获取unix时间与uptime的时间差
+    unix_time = bpf_get_ktime_ns()*1e-9+delta
+
+    Returns:
+        _type_: _description_
+    """
+    with open("/proc/uptime", "r") as f:
+        uptime = float(f.readline().split(" ")[0])
+    delta = time.time() - uptime   # delta是uptime和unix epoch time的差值，因为ebpf虚拟机只能获取uptime所以在前端重新转换为unix epoch time
+    return delta
